fix(electron): use dynamic Windows deny prefixes in file-ipc.js (#1332)
* docs: fix broken link to storage-system.md in linting-rules.md (#1313)

* Initial plan

* docs: fix broken link in linting-rules.md

Agent-Logs-Url: https://github.com/Iktahana/illusions/sessions/e0a84e8c-49d4-4c55-aab6-8e0e7ce999a2


---------


* docs: update docs/ to reflect recent code changes (#1312)

* Initial plan

* docs: update export-system.md and milkdown-plugin.md to reflect TXT blank-line fix and onNlpError callback

Agent-Logs-Url: https://github.com/Iktahana/illusions/sessions/ba351d60-0010-47ee-890f-3a5cedc229a5


---------


* fix(electron): use dynamic Windows deny prefixes in file-ipc.js (#1327)

Extract getWindowsDenyPrefixes() from vfs-ipc.js into path-utils.js as
a shared utility. Replace hardcoded C: drive prefixes in file-ipc.js
with the dynamic function that reads process.env.SystemRoot. Also adds
the missing ProgramData directory to file-ipc.js's deny list.


---------
